Create module to create patch function
